pyproject.toml
# =====================
# Sezione: Sistema di Build
# Specifica i requisiti e il backend per la build del progetto
# =====================
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# =====================
# Sezione: Progetto
# Metadati principali del progetto Python
# =====================
[project]
name = "license-tool"
version = "0.1.0"
description = "Tool for software license analysis and compatibility"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Riccio Giuseppe, Simeone Lucia, Medugno Vittoria, Capone Antonella, Liparulo Elisa"},
]
# Correzione avviso licenza: utilizzo del campo 'license' invece del classifier
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi~=0.121.1",
    "uvicorn[standard]",
    "gitpython~=3.1.45",
    "python-dotenv~=1.2.1",
    "requests~=2.32.5",
    "pydantic~=2.12.4",
    "httpx~=0.28.1",
    "python-multipart~=0.0.20",
    "license-expression"
]

# =====================
# Sezione: URL del Progetto
# Link utili come homepage e bug tracker
# =====================
[project.urls]
"Homepage" = "https://github.com/giusk10/license_tool"
"Bug Tracker" = "https://github.com/giusk10/license_tool/issues"

# =====================
# Sezione: Dipendenze Opzionali
# Dipendenze utili per lo sviluppo e il testing
# =====================
[project.optional-dependencies]
dev = [
    "pytest~=9.0.1",
    "pytest-cov~=7.0.0",
    "pytest-mock~=3.15.1",
    "pytest-asyncio~=1.3.0",
    "pylint",
    "pre-commit",
    "behave~=1.3.3",
    "selenium~=4.39.0",
    "webdriver-manager~=4.0.2"
]

# =====================
# Configurazione setuptools: include solo la cartella 'app', esclude 'frontend' e 'tests'
# =====================
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["frontend*", "tests*"]

# =====================
# Configurazione Pytest
# Opzioni e percorsi per i test automatici
# =====================
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --cov=app"
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = "."

# =====================
# Configurazione Pylint
# Analisi statica del codice e controllo dei messaggi
# =====================
[tool.pylint.main]
# Imposta la root del progetto come sorgente
source-roots = ["."]
# Hook per includere la directory corrente
init-hook = "import sys; sys.path.append('.')"
# Ignora la cartella dei test
ignore = ["tests"]

[tool.pylint."MESSAGES CONTROL"]
disable = ["all"]
enable = ["E", "F"]
# Abilita solo errori e fatal error
